# Optimize the HELLO command reply