@sundb 2024-08-27
`HELLO` command is not a hot path, it it worth to optimize it by reducing some readability.