@fcostaoliveira 2024-08-27
### CE Performance Automation : step 1 of 2 (build) STARTING...

This comment was automatically generated given a benchmark was triggered.
Started building at 2024-08-27 15:18:00.511478
You can check each build/benchmark progress in grafana:
   - git hash: a88f45aed24bb9dba0421e6d23935ceeaa051f4d
   - git branch: filipecosta90:optimize.hello
   - commit date and time: n/a
   - commit summary: n/a
   - test filters:
       - command priority lower limit: 0
       - command priority upper limit: 10000
       - test name regex: .*
       - command group regex: .*