@fcostaoliveira 2024-08-27
### CE Performance Automation : step 1 of 2 (build) DONE.

This comment was automatically generated given a benchmark was triggered.
Started building at 2024-08-27 15:18:00.511478 and took 86 seconds.
You can check each build/benchmark progress in grafana:
   - git hash: a88f45aed24bb9dba0421e6d23935ceeaa051f4d
   - git branch: filipecosta90:optimize.hello
   - commit date and time: n/a
   - commit summary: n/a
   - test filters:
       - command priority lower limit: 0
       - command priority upper limit: 10000
       - test name regex: .*
       - command group regex: .*

You can check a comparison in detail via the [grafana link](https://benchmarksredisio.grafana.net/d/edsxdsrbexhc0f/ce-benchmark-run-status?orgId=1&var-benchmark_work_stream=1724771967463-0)