@fcostaoliveira 2024-08-27
### Automated performance analysis summary

This comment was automatically generated given there is performance data available.

Using platform named: intel64-ubuntu22.04-redis-icx1 to do the comparison.

In summary:
- Detected a total of 1 improvements above the improvement water line.

You can check a comparison in detail via the [grafana link](https://benchmarksredisio.grafana.net/d/1fWbtb7nz/experimental-oss-spec-benchmarks/?var-branch=unstable&var-branch=optimize.hello)

### Comparison between unstable and optimize.hello.

Time Period from 5 months ago. (environment used: oss-standalone)

#### Improvements Table

|                                                                                          Test Case                                                                                          |Baseline unstable (median obs. +- std.dev)|Comparison optimize.hello (median obs. +- std.dev)|% change (higher-better)|   Note    |
|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|-----------------------------------------:|-------------------------------------------------:|------------------------|-----------|
|[memtier_benchmark-connection-hello](https://github.com/redis/redis-benchmarks-specification/blob/v0.1.208/redis_benchmarks_specification/test-suites/memtier_benchmark-connection-hello.yml)|                                    169883|                                            174951|3.0%                    |IMPROVEMENT|


Improvements test regexp names: memtier_benchmark-connection-hello

<details>
  <summary>Full Results table:</summary>

|                                                                                          Test Case                                                                                          |Baseline unstable (median obs. +- std.dev)|Comparison optimize.hello (median obs. +- std.dev)|% change (higher-better)|   Note    |
|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|-----------------------------------------:|-------------------------------------------------:|------------------------|-----------|
|[memtier_benchmark-connection-hello](https://github.com/redis/redis-benchmarks-specification/blob/v0.1.208/redis_benchmarks_specification/test-suites/memtier_benchmark-connection-hello.yml)|                                    169883|                                            174951|3.0%                    |IMPROVEMENT|

</details>