@fcostaoliveira 2024-09-02
### CE Performance Automation : step 1 of 2 (build) DONE.

This comment was automatically generated given a benchmark was triggered.
Started building at 2024-09-02 13:57:27.152553 and took 88 seconds.
You can check each build/benchmark progress in grafana:
   - git hash: c8b3817506964ce6dbc4c12f0d4f58775098cd4e
   - git branch: filipecosta90:optimize.hello
   - commit date and time: n/a
   - commit summary: n/a
   - test filters:
       - command priority lower limit: 0
       - command priority upper limit: 10000
       - test name regex: .*
       - command group regex: .*

You can check a comparison in detail via the [grafana link](https://benchmarksredisio.grafana.net/d/edsxdsrbexhc0f/ce-benchmark-run-status?orgId=1&var-benchmark_work_stream=1725285536305-0)